fix the bug of float to int conversion

+++ calculator.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int char_to_int(char x);
-int number(char [], int &sindex); // this function will make numbers from the evaluate string
-int mult_divide(char [], int &index);
+float number(char [], int &sindex); // this function will make numbers from the evaluate string
+float mult_divide(char [], int &index);
 
 int main()
 {
@@ -13,7 +13,7 @@ int main()
     cout << "Enter expression to evaluate: ";
     cin >> eval_string;
     
-    int ans = 0; // Final answer will be stored here
+    float ans = 0; // Final answer will be stored here
     for (int i =0;  eval_string[i] != '\0';)
     {
       if (eval_string[i] == '-')
@@ -42,9 +42,9 @@ int char_to_int(char x)
 }
 
 // this function will make the number from character like if 120 are three characters '1','2','0' then this function will return a number 120
-int number( char str[], int &index)
+float number( char str[], int &index)
 {
-  int n= 0;
+  float n= 0;
   while (str[index] >= '0' && str[index] <= '9')
   {
     n = n* 10 + char_to_int(str[index]);
@@ -54,9 +54,9 @@ int number( char str[], int &index)
 }
 
 // This function will perform the calculation for multiplication and divsion and return the answer
-int mult_divide(char num[], int &index)
+float mult_divide(char num[], int &index)
 {
-    int temp = number(num, index); // this will hold the first the number
+    float temp = number(num, index); // this will hold the first the number
     while (num[index] != '\0' && num[index] != '+' && num[index] != '-')
     {
       if ( num[index] == '*')
